make sshp: s/[[user]@]/[user@]/
I used to have a version that w/o user@ prefix before hostname
`nemo@` was prepended to the hostname, and w/ plain '@' just the
hostname following @ was used.

The version in this Makefile never had that feature.

The example shows `nemo@192.168.2.15` as this could have used to
access device before one used mad-developer to start sshd -- and
it still works done so (even to access defaultuser@... devices).